test-demo/requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
